feat(protocols): enhance MQTTv5 protocol with extended Last Will and Testament (LWT) support
- Added setters for configurable LWT properties in `Connect5` and `MQTTPacket5`.
- Modified `MQTT5ProtocolFactory` to support dynamic LWT message construction via `MqttWillConfigDTO`.
- Introduced `WillDelayInterval`, `ContentType`, and `MessageExpiryInterval` handling in LWT properties.
- Adjusted `MqttWillConfig` to use `int` for `delayInterval` instead of `long`.
- Extended `MQTT5Protocol` to pass LWT configuration during session connection.

+++ src/main/java/io/mapsmessaging/config/network/MqttWillConfig.java
@@ -38,7 +38,7 @@ public class MqttWillConfig extends MqttWillConfigDTO implements Config {
     this.payloadEncoding = properties.getProperty("payloadEncoding", payloadEncoding);
     this.qos = properties.getIntProperty("qos", qos);
     this.retain = properties.getBooleanProperty("retain", retain);
-    this.delayInterval = properties.getLongProperty("delayInterval", delayInterval);
+    this.delayInterval = properties.getIntProperty("delayInterval", delayInterval);
     this.messageExpiryInterval = properties.getLongProperty("messageExpiryInterval", messageExpiryInterval);
     this.contentType = properties.getProperty("contentType", contentType);
     this.payloadFormatIndicator = properties.getIntProperty("payloadFormatIndicator", payloadFormatIndicator);
